feat: 체스판 객체 구현, 기물 이동 및 점수 계산 기능 재배치, 기능 수행 주체 변경에 따른 테스트 수정
