+++ src/main/java/chess/domain/board/PieceIdentifier.java
@@ -1,0 +1,9 @@
+package chess.domain.board;
+
+import chess.domain.piece.Piece;
+
+@FunctionalInterface
+public interface PieceIdentifier {
+
+    boolean identify(Piece piece);
+}
