+++ src/main/java/chess/domain/state/BlackWin.java
@@ -1,12 +1,11 @@
 package chess.domain.state;
 
-import chess.domain.board.Rank;
-import java.util.Map;
+import chess.domain.board.Board;
 
 public final class BlackWin extends End {
 
-    public BlackWin(Map<Integer, Rank> ranks) {
-        super(ranks);
+    public BlackWin(Board board) {
+        super(board);
     }
 
     @Override
