+++ src/main/java/chess/domain/state/GameState.java
@@ -1,6 +1,6 @@
 package chess.domain.state;
 
-import chess.domain.board.Rank;
+import chess.domain.board.Board;
 import chess.domain.piece.Position;
 
 public interface GameState {
@@ -13,17 +13,9 @@ public interface GameState {
 
     GameState move(Position start, Position target);
 
-    Rank getRank(int rankLine);
-
     GameState terminate();
 
-    double calculateBlackScore();
-
-    double calculateWhiteScore();
-
-    End judgeWinner();
-
-    Playing judgeTurn();
-
     String findTurn();
+
+    Board getBoard();
 }
