+++ src/main/java/chess/domain/state/Terminate.java
@@ -1,12 +1,11 @@
 package chess.domain.state;
 
-import chess.domain.board.Rank;
-import java.util.Map;
+import chess.domain.board.Board;
 
 public final class Terminate extends End {
 
-    public Terminate(Map<Integer, Rank> ranks) {
-        super(ranks);
+    public Terminate(Board board) {
+        super(board);
     }
 
     @Override
