+++ src/main/java/chess/domain/state/WhiteWin.java
@@ -1,12 +1,11 @@
 package chess.domain.state;
 
-import chess.domain.board.Rank;
-import java.util.Map;
+import chess.domain.board.Board;
 
 public final class WhiteWin extends End {
 
-    public WhiteWin(Map<Integer, Rank> ranks) {
-        super(ranks);
+    public WhiteWin(Board board) {
+        super(board);
     }
 
     @Override
