+++ src/test/java/chess/domain/board/BoardInitializerTest.java
@@ -1,9 +1,8 @@
-package chess.domain.board.state;
+package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.piece.Piece;
-import chess.domain.state.BoardInitializer;
 import chess.domain.state.GameState;
 import chess.domain.state.WhiteTurn;
 import java.util.List;
@@ -16,7 +15,7 @@ class BoardInitializerTest {
     @DisplayName("체스판 초기화 시 모든 기물이 제자리에 생성되었는지 확인")
     @Test
     void initializeBoard() {
-        GameState boardState = BoardInitializer.initBoard();
+        GameState gameState = new WhiteTurn(new Board(BoardInitializer.initBoard()));
 
         List<List<String>> signatures = List.of(
                 List.of("R", "N", "B", "Q", "K", "B", "N", "R"),
@@ -30,7 +29,7 @@ class BoardInitializerTest {
         );
 
         for (int i = 0; i < 8; i++) {
-            assertThat(boardState.getRank(7 - i).getPieces().stream()
+            assertThat(gameState.getBoard().getRank(7 - i).getPieces().stream()
                     .map(Piece::getSignature)
                     .collect(Collectors.toList())
             ).isEqualTo(signatures.get(i));
@@ -40,7 +39,7 @@ class BoardInitializerTest {
     @DisplayName("체스 게임은 백팀 차례로 시작한다.")
     @Test
     void isWhiteTurnAfterInitialize() {
-        GameState whiteTurn = BoardInitializer.initBoard();
+        GameState whiteTurn = new WhiteTurn(new Board(BoardInitializer.initBoard()));
 
         assertThat(whiteTurn).isInstanceOf(WhiteTurn.class);
     }
