--- src/test/java/chess/domain/board/state/PlayingTest.java
@@ -1,72 +1,0 @@
-package chess.domain.board.state;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-import chess.domain.piece.Position;
-import chess.domain.state.BoardInitializer;
-import chess.domain.state.GameState;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-public class PlayingTest {
-
-    @DisplayName("대각선 직선 이동 시 중간에 기물이 존재한다면 이동할 수 없다.")
-    @Test
-    void diagonalStraightMove_Obstacle_Fails() {
-        GameState boardState = BoardInitializer.initBoard();
-
-        assertThatThrownBy(() -> boardState.move(new Position("c1"), new Position("a3")))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("경로에 기물이 존재합니다.");
-    }
-
-    @DisplayName("대각선으로 이동하려는 위치에 아군 말이 존재하면 예외가 발생한다.")
-    @Test
-    void diagonalStraightMove_ObstacleAlly_Fails() {
-        GameState boardState = BoardInitializer.initBoard();
-
-        assertThatThrownBy(() -> boardState.move(new Position("c1"), new Position("b2")))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이동이 불가능한 위치입니다.");
-    }
-
-    @DisplayName("세로로 직선 이동 시 중간에 기물이 존재한다면 이동할 수 없다.")
-    @Test
-    void verticalStraightMove_Obstacle_Fails() {
-        GameState boardState = BoardInitializer.initBoard();
-
-        assertThatThrownBy(() -> boardState.move(new Position("a1"), new Position("a3")))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("경로에 기물이 존재합니다.");
-    }
-
-    @DisplayName("세로로 이동하려는 위치에 아군 말이 존재하면 예외가 발생한다.")
-    @Test
-    void verticalStraightMove_ObstacleAlly_Fails() {
-        GameState boardState = BoardInitializer.initBoard();
-
-        assertThatThrownBy(() -> boardState.move(new Position("a1"), new Position("a2")))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이동이 불가능한 위치입니다.");
-    }
-
-    @DisplayName("가로로 직선 이동 시 중간에 기물이 존재한다면 이동할 수 없다.")
-    @Test
-    void horizontalStraightMove_Obstacle_Fails() {
-        GameState boardState = BoardInitializer.initBoard();
-
-        assertThatThrownBy(() -> boardState.move(new Position("a1"), new Position("c1")))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("경로에 기물이 존재합니다.");
-    }
-
-    @DisplayName("가로로 이동하려는 위치에 아군 말이 존재하면 예외가 발생한다.")
-    @Test
-    void horizontalStraightMove_ObstacleAlly_Fails() {
-        GameState boardState = BoardInitializer.initBoard();
-
-        assertThatThrownBy(() -> boardState.move(new Position("a1"), new Position("b1")))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이동이 불가능한 위치입니다.");
-    }
-}
